requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
coverage==6.4.1
Django==4.0.5
django-quill-editor==0.1.40
execnet==1.9.0
frozenlist==1.4.1
idna==3.3
iniconfig==1.1.1
multidict==6.0.5
numpy==1.23.0
packaging==21.3
pandas==1.4.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.9.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-lazy-fixture==0.6.3
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
twilio==9.2.3
tzdata==2022.1
urllib3==1.26.9
yarl==1.9.4
